+++ R/SimDesign.R
@@ -25,6 +25,7 @@
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
 #' @import methods parallel stats
 #' @importFrom sessioninfo session_info
+#' @importFrom parallelly availableCores
 #' @importFrom RPushbullet pbPost
 #' @importFrom graphics abline
 #' @importFrom R.utils withTimeout
@@ -34,6 +35,7 @@
 #' @importFrom future.apply future_lapply
 #' @importFrom progressr progressor
 #' @importFrom beepr beep
+#' @importFrom snow sendCall checkForRemoteErrors
 # @importFrom robustbase glmrob
 #' @importFrom utils recover packageVersion head tail capture.output object.size
 #' @keywords package