+++ R/runArraySimulation.R
@@ -7,7 +7,7 @@
 #' jobs to HPC clusters where a job array number is available (e.g., via SLURM),
 #' where the simulation results must be saved to independent files as they
 #' complete. Use of \code{\link{expandDesign}} is useful for distributing replications
-#' to different jobs, while \code{\link{gen_seeds}} is required to ensure high-quality
+#' to different jobs, while \code{\link{genSeeds}} is required to ensure high-quality
 #' random number generation across the array submissions. See the associated
 #' vignette for a brief tutorial of this setup.
 #'
@@ -17,9 +17,10 @@
 #' approach, which uses this method to distribute random seeds within
 #' each isolated condition rather than between all conditions). As such, this
 #' function requires the seeds to be generated using
-#' \code{\link{gen_seeds}} with the \code{iseed} and \code{arrayID}
+#' \code{\link{genSeeds}} with the \code{iseed} and \code{arrayID}
 #' inputs to ensure that each job is analyzing a high-quality
-#' set of random numbers via L'Ecuyer-CMRG's (2002) method.
+#' set of random numbers via L'Ecuyer-CMRG's (2002) method, incremented using
+#' \code{\link[parallel]{nextRNGStream}}.
 #'
 #' Additionally, for timed simulations on HPC clusters it is also recommended to pass a
 #' \code{control = list(max_time)} value to avoid discarding
@@ -48,22 +49,40 @@
 #'
 #' @param filename file name to save simulation files to (does not need to
 #'   specify extension). However, the array ID will be appended to each
-#'   \code{filename} (see \code{filename_suffix}). For example, if
+#'   \code{filename}. For example, if
 #'   \code{filename = 'mysim'} then files stored will be \code{'mysim-1.rds'},
-#'   \code{'mysim-2.rds'}, and so on for each row in \code{design}
+#'   \code{'mysim-2.rds'}, and so on for each row ID in \code{design}
 #'
 #' @param dirname directory to save the files associated with \code{filename}
 #'   to. If omitted the files will be stored in the same working directory
 #'   where the script was submitted
 #'
-#' @param filename_suffix suffix to add to the \code{filename};
-#'   default add '-' with the \code{arrayID}
+#' @param parallel logical; use parallel computations via the a "SOCK" cluster?
+#'   Only useful when the instruction shell file requires more than 1 core
+#'   (number of cores detected via \code{ncores}). For this application
+#'   the random seeds further distributed using \code{\link[parallel]{nextRNGSubStream}}
 #'
-#' @param iseed initial seed to be passed to \code{\link{gen_seeds}}'s argument
+#' @param cl cluster definition. If omitted a "SOCK" cluster will be defined
+#'
+#' @param ncores number of cores to use when \code{parallel=TRUE}. Note that
+#'   the default uses 1 minus the number of available cores, therefore this
+#'   will only be useful when \code{ncores > 2} as defined in the shell instruction
+#'   file
+#'
+#' @param iseed initial seed to be passed to \code{\link{genSeeds}}'s argument
 #'   of the same name, along with the supplied \code{arrayID}
 #'
 #' @param addArrayInfo logical; should the array ID and original design row number
-#'   be added to the \code{SimExtract(..., what='results')} output?
+#'   be added to the \code{SimResults(...)} output?
+#'
+#' @param array2row user defined function with the single argument \code{arrayID}.
+#'   Used to convert the detected \code{arrayID}
+#'   into a suitable row index in the \code{design} object input. By default
+#'   each \code{arrayID} is associated with its respective row in \code{design}.
+#'
+#'   For example, if each \code{arrayID} should evaluate 10 rows in
+#'   the \code{design} object then the function
+#'   \code{function(arrayID){1:10 + 10 * (arrayID-1)}} can be passed to \code{array2row}
 #'
 #' @param control control list passed to \code{\link{runSimulation}}.
 #'   In addition to the original \code{control} elements two
@@ -102,12 +121,9 @@
 #' with the SimDesign Package. \code{The Quantitative Methods for Psychology, 16}(4), 248-280.
 #' \doi{10.20982/tqmp.16.4.p248}
 #'
-#' Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
-#' Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
-#' \doi{10.1080/10691898.2016.1246953}
-#'
 #' @seealso \code{\link{runSimulation}}, \code{\link{expandDesign}},
-#'   \code{\link{gen_seeds}}, \code{\link{SimCollect}}, \code{\link{getArrayID}}
+#'   \code{\link{genSeeds}}, \code{\link{SimCheck}},
+#'   \code{\link{SimCollect}}, \code{\link{getArrayID}}
 #'
 #' @examples
 #'
@@ -139,7 +155,8 @@
 #' ###   and therefore should be used in real SLURM submissions
 #' arrayID <- getArrayID(type = 'slurm')
 #'
-#' # However, for the following example array ID is set to first row only
+#' # However, the following example arrayID is set to
+#' #  the first row only for testing purposes
 #' arrayID <- 1L
 #'
 #' # run the simulation (results not caught on job submission, only files saved)
@@ -150,8 +167,17 @@
 #' res
 #' SimResults(res) # condition and replication count stored
 #'
+#' # same, but evaluated with multiple cores
+#' res <- runArraySimulation(design=Design, replications=50,
+#'                       generate=Generate, analyse=Analyse,
+#'                       summarise=Summarise, arrayID=arrayID,
+#'                       parallel=TRUE, ncores=3,
+#'                       iseed=iseed, filename='myparsim')
+#' res
+#' SimResults(res) # condition and replication count stored
+#'
 #' dir()
-#' SimClean('mysim-1.rds')
+#' SimClean(c('mysim-1.rds', 'myparsim-1.rds'))
 #'
 #' ########################
 #' # Same submission job as above, however split the replications over multiple
@@ -180,6 +206,8 @@
 #'
 #' ###
 #' # Emulate the arrayID distribution, storing all results in a 'sim/' folder
+#' # (if 'sim/' does not exist in runArraySimulation() it will be
+#' # created automatically)
 #' dir.create('sim/')
 #'
 #' # Emulate distribution to nrow(Design5) = 15 independent job arrays
@@ -199,18 +227,57 @@
 #' # list saved files
 #' dir('sim/')
 #'
-#' setwd('sim')
-#' condition14 <- readRDS('condition-14.rds')
+#' # check that all files saved (warnings will be raised if missing files)
+#' SimCheck('sim/') |> isTRUE()
+#'
+#' condition14 <- readRDS('sim/condition-14.rds')
 #' condition14
 #' SimResults(condition14)
 #'
 #' # aggregate simulation results into single file
-#' final <- SimCollect(files=dir())
+#' final <- SimCollect('sim/')
 #' final
 #'
-#' SimResults(final) |> View()
+#' # clean simulation directory
+#' SimClean(dirs='sim/')
 #'
-#' setwd('..')
+#'
+#' ############
+#' # same as above, however passing different amounts of information depending
+#' # on the array ID
+#' array2row <- function(arrayID){
+#'   switch(arrayID,
+#'     "1"=1:8,
+#'     "2"=9:14,
+#'     "3"=15)
+#' }
+#'
+#' # arrayID 1 does row 1 though 8, arrayID 2 does 9 to 14
+#' array2row(1)
+#' array2row(2)
+#' array2row(3)  # arrayID 3 does 15 only
+#'
+#' # emulate remote array distribution with only 3 arrays
+#' sapply(1:3, \(arrayID)
+#'      runArraySimulation(design=Design5, replications=10,
+#'           generate=Generate, analyse=Analyse,
+#'           summarise=Summarise, iseed=iseed, arrayID=arrayID,
+#'           filename='condition', dirname='sim', array2row=array2row)) |> invisible()
+#'
+#' # list saved files
+#' dir('sim/')
+#'
+#' # note that all row conditions are still stored separately, though note that
+#' #  arrayID is now 2 instead
+#' condition14 <- readRDS('sim/condition-14.rds')
+#' condition14
+#' SimResults(condition14)
+#'
+#' # aggregate simulation results into single file
+#' final <- SimCollect('sim/')
+#' final
+#'
+#' # clean simulation directory
 #' SimClean(dirs='sim/')
 #'
 #' }
@@ -218,54 +285,75 @@
 runArraySimulation <- function(design, ..., replications,
                                iseed, filename, dirname = NULL,
                                arrayID = getArrayID(),
-                               filename_suffix = paste0("-", arrayID),
+                               array2row = function(arrayID) arrayID,
                                addArrayInfo = TRUE,
+                               parallel = FALSE, cl = NULL,
+                               ncores = parallelly::availableCores(omit = 1L),
                                save_details = list(),
                                control = list()){
     dots <- list(...)
+    if(parallel && ncores == 1L) parallel <- FALSE
     if(!is.null(dots$save_results) && isTRUE(dots$save_results))
         stop('save_results not supported for array jobs. Please use store_results only')
     if(!is.null(control$save_seeds) && isTRUE(control$save_seeds))
         stop(c('save_seeds not supported for array jobs. If this is truely',
                ' necessary use store_Random.seeds instead'))
+    control$from.runArraySimulation <- TRUE
     rngkind <- RNGkind()
     RNGkind("L'Ecuyer-CMRG")
     on.exit(RNGkind(rngkind[1L]))
     stopifnot(!missing(design))
-    if(is.null(attr(design, 'condition')))
-        attr(design, 'condition') <- 1L:nrow(design)
+    if(is.null(attr(design, 'Design.ID')))
+        attr(design, 'Design.ID') <- 1L:nrow(design)
     stopifnot(!missing(iseed))
     stopifnot(!missing(filename))
+    if(grepl('-', filename))
+        stop('filename cannot contain the character \"-\" as this is used in the suffix')
     stopifnot(nrow(design) > 1L)
-    stopifnot(!missing(replications))
-    if(length(replications) == 1L)
-        replications <- rep(replications, nrow(design))
-    stopifnot(length(replications) == nrow(design))
     stopifnot("arrayID is not a single integer identifier"=
                   length(arrayID) == 1L && is.numeric(arrayID) && !is.na(arrayID))
-    stopifnot(arrayID %in% 1L:nrow(design))
+    rowpick <- array2row(arrayID)
+    filename_suffix <- paste0("-", rowpick)
+    stopifnot(!missing(replications))
+    if(length(replications) > 1L)
+        replications <- replications[rowpick]
+    stopifnot(rowpick %in% 1L:nrow(design))
     if(!is.null(filename))
         filename <- paste0(filename, filename_suffix)
     if(!is.null(dirname)){
+        if(!dir.exists(dirname))
+            dir.create(dirname, showWarnings = FALSE)
         filename <- file.path(dirname, filename)
         filename <- gsub("//", "/", filename)
     }
     save_details$arrayID <- arrayID
-    seed <- genSeeds(design, iseed=iseed, arrayID=arrayID)
-
-    ret <- runSimulation(design=design[arrayID, , drop=FALSE],
-                         replications=replications[arrayID],
-                         filename=filename, seed=seed,
-                         verbose=FALSE, save_details=save_details,
-                         control=control, ...)
-    if(addArrayInfo && (is.null(dots$store_results) ||
-       (!is.null(dots$store_results) && isTRUE(dots$store_results)))){
-        results <- SimExtract(ret, 'results')
-        condition <- attr(design, 'condition')
-        results <- dplyr::mutate(results, arrayID=arrayID, .before=1L)
-        results <- dplyr::mutate(results, condition=condition[arrayID], .before=1L)
-        attr(ret, "extra_info")$stored_results <- results
-        saveRDS(ret, paste0(filename, '.rds'))
+    if(parallel){
+        if(is.null(cl)){
+            cl <- parallel::makeCluster(ncores, type="SOCK")
+            on.exit(parallel::stopCluster(cl), add=TRUE)
+        }
     }
+    for(i in 1L:length(rowpick)){
+        row <- rowpick[i]
+        seed <- genSeeds(design, iseed=iseed, arrayID=row)
+        dsub <- design[row, , drop=FALSE]
+        attr(dsub, 'Design.ID') <- attr(design, 'Design.ID')[row]
+        ret <- runSimulation(design=dsub, replications=replications, seed=seed,
+                             verbose=FALSE, save_details=save_details,
+                             parallel=parallel, cl=cl,
+                             control=control, save=FALSE, ...)
+        attr(ret, 'extra_info')$number_of_conditions <- nrow(design)
+        if(addArrayInfo && (is.null(dots$store_results) ||
+           (!is.null(dots$store_results) && isTRUE(dots$store_results)))){
+            results <- SimExtract(ret, 'results')
+            condition <- attr(design, 'Design.ID')
+            results <- dplyr::mutate(results, arrayID=arrayID, .before=1L)
+            results <- dplyr::mutate(results, condition=condition[row], .before=1L)
+            attr(ret, "extra_info")$stored_results <- results
+        }
+        filename.u <- unique_filename(filename[i], safe=TRUE, verbose=FALSE)
+        saveRDS(ret, filename.u)
+    }
+    if(length(rowpick) > 1L) ret <- NULL
     invisible(ret)
 }